Refactor CreateAsynchronousMailController script
Utilize existing scripts from grails-core to create artifacts from templates.
This also provides the benefit of allowing users to select an artifact name
of their choosing. If no name is provided, a default name of
'AsynchronousMailController' will be used. The script creates an associated
unit test for the controller as well.